Added piecewise linear arrival and initial arrivals

+++ Event.java
@@ -2,18 +2,18 @@ import java.util.Comparator;
 
 public class Event {
 	private EventType type;
-	private float time;
+	private double time;
 	private Elevator elevator;
 	private Person person;
 
-	Event(EventType type, float time, Person person) {
+	Event(EventType type, double time, Person person) {
 		this.type = type;
 		this.time = time;
 		this.person = person;
 		this.elevator = null;
 	}
 
-	Event(EventType type, float time, Elevator elevator) {
+	Event(EventType type, double time, Elevator elevator) {
 		this.type = type;
 		this.time = time;
 		this.elevator = elevator;
@@ -24,7 +24,7 @@ public class Event {
 		return type;
 	}
 
-	float getTime() {
+	double getTime() {
 		return time;
 	}
